ready cyrrilic fonts

+++ matr_mini/display.h
@@ -23,7 +23,7 @@ class Display : public MATR{
   protected:
     uint8_t reverseChar(uint8_t ch);
     uint8_t *getAsciiArray(uint8_t ch);
-    uint8_t *getCyrillicArray(uint8_t ch);
+    uint8_t *getCyrillicArray(unsigned short int ch);
     void insertToDispBuffer(uint8_t *array);
     void insertToDispBuffer(uint8_t *array, int sh);
